% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fit_hmd.R
\name{glm_fit_hmd}
\alias{glm_fit_hmd}
\title{GLM model}
\usage{
glm_fit_hmd(
  data = aiinsurance::insurance_train,
  target = "outcome",
  family = "binomial"
)
}
\arguments{
\item{data}{An arbitrary dataframe
except the target}

\item{target}{The target variable aimed for prediction}

\item{family}{specify family of distribution.
The options are binomial, gaussian Gamma, inverse.gaussian, poisson, quasi,
quasibinomial, quasipoisson}
}
\value{
Returns fit object of glm function
}
\description{
GLM model
}
\details{
This functions allows the user to perform generalized linear model (GLM) on a given dataframe by specifying feature names (response variables),
target variable, family of distribution, and the dataset.
}
