% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_predict_hmd.R
\name{glm_predict_hmd}
\alias{glm_predict_hmd}
\title{Predict data with results using GLM fit object}
\usage{
glm_predict_hmd(
  fit,
  data = aiinsurance::insurance_test,
  target = "outcome",
  threshold = 0.5
)
}
\arguments{
\item{fit}{fit object from a glm model}

\item{data}{An arbitrary dataframe}

\item{target}{Target column in the data}

\item{threshold}{Cutoff point of probability of class 1, to classify the inputs
with the probability above threshold as 1.}
}
\value{
Returns fit object of glm function
}
\description{
Predict data with results using GLM fit object
}
\details{
This functions allows the user to predict the results obtained from GLM model
}
