% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_cv_predict_hmd.R
\name{glmnet_cv_predict_hmd}
\alias{glmnet_cv_predict_hmd}
\title{Predict data with results from CV.GLMNET model}
\usage{
glmnet_cv_predict_hmd(
  fit,
  data = aiinsurance::insurance_test,
  target = "outcome",
  lchoice = "min",
  threshold = 0.5
)
}
\arguments{
\item{fit}{fit object from a cv.glmnet model}

\item{data}{An arbitrary dataframe}

\item{target}{Target column in the data}

\item{lchoice}{how to choose lambda model. Choices are "min", "1se".}

\item{threshold}{Cutoff point of probability of class 1, to classify the inputs
with the probability above threshold as 1.}
}
\value{
Returns a hash of predictions, prediction probabilities, and coefficients
}
\description{
Predict data with results from CV.GLMNET model
}
\details{
This functions allows the user to predict a given dataset using fit object
obtained from the cv.glmnet function
}
