% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_splitter_hmd.R
\name{train_test_splitter_hmd}
\alias{train_test_splitter_hmd}
\title{Splitting a Dataframe to Training and Test Datasets}
\usage{
train_test_splitter_hmd(
  data = aiinsurance::car_insurance_data,
  proportion = 0.7,
  join = "ID"
)
}
\arguments{
\item{data}{An arbitrary dataframe}

\item{proportion}{Proportion of train to test. Default is 0.7}

\item{join}{The common column of train and test, by which separation is based on.}
}
\value{
Returns a hash containing train and test data
}
\description{
Splitting a Dataframe to Training and Test Datasets
}
\details{
This functions allows the user to split data into train and test. The function
needs a common column in both train and test, on which the separation is based.
In car_insurance_data, the join column is ID
}
\examples{
h <- train_test_splitter_hmd(car_insurance_data, proportion=0.7, join="ID")
train <- h$train
test <- h$test
dplyr::glimpse(train)
dplyr::glimpse(test)
}
